Make Chord Diagram

Makes a chord diagram representing the frequency of interactions between features

Usage

plot_chord(
  fit,
  featureNames,
  featureGroups,
  linkColors,
  gridColors,
  bgLinkColor = "gray",
  maxLen,
  lengths = c(1:maxLen),
  dims = c(1, maxLen),
  minProp = 0,
  textSize = 1,
  line_arg = 1,
  side_mar = 0,
  top_mar = 0
)

Arguments

fit

output from function BRS

featureNames

names of features as they appeared in df

featureGroups

dataframe whose first column is the features (only the prefixes, not the values) in featureNames and whose second column is the corresponding labels to display

linkColors

vector of colors to use for chords

gridColors

vector of colors to use for arcs

minProp

minimum proportion of times a relationship must appear to be plotted

textSize

cex parameter to be passed to par

Value

chord diagram